+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Fixed
 ### Changed
 - Dependencies - Updated Microsoft.CodeAnalysis.CSharp to 5.3.0
+- Dependencies - Updated FunFair.CodeAnalysis to 7.1.35.1745
 ### Removed
 ### Deployment Changes