+++ CHANGELOG.md
@@ -10,8 +10,8 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Fixed
 ### Changed
 - Dependencies - Updated FunFair.CodeAnalysis to 7.1.36.1750
-- Dependencies - Updated Meziantou.Analyzer to 3.0.28
 - Dependencies - Updated SonarAnalyzer.CSharp to 10.22.0.136894
+- Dependencies - Updated Meziantou.Analyzer to 3.0.46
 ### Removed
 ### Deployment Changes