+++ CHANGELOG.md
@@ -10,10 +10,10 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Fixed
 ### Changed
 - Dependencies - Updated FunFair.CodeAnalysis to 7.1.36.1750
-- Dependencies - Updated SonarAnalyzer.CSharp to 10.22.0.136894
 - Dependencies - Updated Meziantou.Analyzer to 3.0.46
 - Dependencies - Updated Philips.CodeAnalysis.DuplicateCodeAnalyzer to 1.11.1
 - Dependencies - Updated Philips.CodeAnalysis.MaintainabilityAnalyzers to 1.11.0
+- Dependencies - Updated SonarAnalyzer.CSharp to 10.23.0.137933
 ### Removed
 ### Deployment Changes