+++ CHANGELOG.md
@@ -10,12 +10,12 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Fixed
 ### Changed
 - Dependencies - Updated FunFair.CodeAnalysis to 7.1.36.1750
-- Dependencies - Updated Philips.CodeAnalysis.MaintainabilityAnalyzers to 1.11.0
 - Dependencies - Updated SonarAnalyzer.CSharp to 10.23.0.137933
 - Dependencies - Updated FunFair.Test.Common to 6.2.19.2017
 - Dependencies - Updated FunFair.Test.Source.Generator to 6.2.19.2017
 - Dependencies - Updated Meziantou.Analyzer to 3.0.48
 - Dependencies - Updated Philips.CodeAnalysis.DuplicateCodeAnalyzer to 1.12.0
+- Dependencies - Updated Philips.CodeAnalysis.MaintainabilityAnalyzers to 1.12.0
 ### Removed
 ### Deployment Changes