+++ CHANGELOG.md
@@ -14,8 +14,8 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 - Dependencies - Updated FunFair.Test.Common to 6.2.19.2017
 - Dependencies - Updated FunFair.Test.Source.Generator to 6.2.19.2017
 - Dependencies - Updated Meziantou.Analyzer to 3.0.48
-- Dependencies - Updated Philips.CodeAnalysis.MaintainabilityAnalyzers to 1.12.0
 - Dependencies - Updated Philips.CodeAnalysis.DuplicateCodeAnalyzer to 2.0.0
+- Dependencies - Updated Philips.CodeAnalysis.MaintainabilityAnalyzers to 2.0.0
 ### Removed
 ### Deployment Changes