+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 - Dependencies - Updated Philips.CodeAnalysis.DuplicateCodeAnalyzer to 2.0.0
 - Dependencies - Updated Philips.CodeAnalysis.MaintainabilityAnalyzers to 2.0.0
 - Dependencies - Updated Meziantou.Analyzer to 3.0.50
+- Dependencies - Updated Credfeto.Version.Information.Generator to 1.0.123.1157
 ### Removed
 ### Deployment Changes