+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Added
 ### Fixed
 ### Changed
+- Dependencies - Updated Meziantou.Analyzer to 3.0.54
 ### Removed
 ### Deployment Changes