+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Fixed
 ### Changed
 - Dependencies - Updated Meziantou.Analyzer to 3.0.54
+- Dependencies - Updated Credfeto.Version.Information.Generator to 1.0.124.1183
 ### Removed
 ### Deployment Changes