+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Changed
 - Dependencies - Updated Meziantou.Analyzer to 3.0.54
 - Dependencies - Updated Credfeto.Version.Information.Generator to 1.0.124.1183
+- Dependencies - Updated FunFair.CodeAnalysis to 7.1.41.1934
 ### Removed
 ### Deployment Changes