+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 - Dependencies - Updated Meziantou.Analyzer to 3.0.58
 - Dependencies - Updated SonarAnalyzer.CSharp to 10.25.0.139117
 - Dependencies - Updated FunFair.Test.Common to 6.2.22.2198
+- Dependencies - Updated FunFair.Test.Source.Generator to 6.2.22.2198
 ### Removed
 ### Deployment Changes