+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Fixed
 ### Changed
 - Dependencies - Updated FunFair.CodeAnalysis to 7.1.42.1940
+- Dependencies - Updated Meziantou.Analyzer to 3.0.74
 ### Removed
 ### Deployment Changes