+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 - Dependencies - Updated FunFair.CodeAnalysis to 7.1.42.1940
 - Dependencies - Updated Meziantou.Analyzer to 3.0.74
 - Dependencies - Updated FunFair.Test.Common to 6.2.23.2204
+- Dependencies - Updated FunFair.Test.Source.Generator to 6.2.23.2204
 ### Removed
 ### Deployment Changes