+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 - Dependencies - Updated Meziantou.Analyzer to 3.0.74
 - Dependencies - Updated FunFair.Test.Common to 6.2.23.2204
 - Dependencies - Updated FunFair.Test.Source.Generator to 6.2.23.2204
+- Dependencies - Updated Credfeto.Version.Information.Generator to 1.0.125.1199
 ### Deprecated
 ### Removed
 ### Deployment Changes