+++ CHANGELOG.md
@@ -12,10 +12,10 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Fixed
 ### Changed
 - Dependencies - Updated FunFair.CodeAnalysis to 7.1.42.1940
-- Dependencies - Updated Meziantou.Analyzer to 3.0.74
 - Dependencies - Updated FunFair.Test.Common to 6.2.23.2204
 - Dependencies - Updated FunFair.Test.Source.Generator to 6.2.23.2204
 - Dependencies - Updated Credfeto.Version.Information.Generator to 1.0.125.1199
+- Dependencies - Updated Meziantou.Analyzer to 3.0.85
 ### Deprecated
 ### Removed
 ### Deployment Changes