+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Added
 ### Fixed
 ### Changed
+- Dependencies - Updated Meziantou.Analyzer to 3.0.86
 ### Deprecated
 ### Removed
 ### Deployment Changes