+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 - Dependencies - Updated FunFair.Test.Common to 6.2.25.2243
 - Dependencies - Updated FunFair.Test.Source.Generator to 6.2.25.2243
 - Dependencies - Updated SonarAnalyzer.CSharp to 10.26.0.140279
+- SDK - Updated DotNet SDK to 10.0.300
 ### Deprecated
 ### Removed
 ### Deployment Changes